python-scripts/expense-tracker/requirements.txt
anthropic
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
